refactor: replace all constants to separate files
