+++ src/constants/index.ts
@@ -1,0 +1,2 @@
+export const MAX_FILE_SIZE = 2 * 1024 * 1024;
+export const ALLOWED_TYPES = ['image/png', 'image/jpeg'];
