+++ src/store/selectors.ts
@@ -1,0 +1,4 @@
+import { RootState } from '../store/store';
+
+export const selectCountries = (state: RootState) =>
+  state.createCountriesSlice || [];
